% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addwaterfallPrep.R
\name{addwaterfallPrep}
\alias{addwaterfallPrep}
\title{Prepares additive data for a waterfall plot}
\usage{
addwaterfallPrep(df, gross.report = 100, NTG.report = 1, NTG.eval = 1,
  altparamnames = NULL, output = "all")
}
\arguments{
\item{df}{a dataframe with columns 'params' (character) and 'value' (numeric); column names are immaterial. This is the order dependent value.}

\item{gross.report}{is the reported gross (ex ante) savings for the program, default is 100}

\item{NTG.report}{is the percentage of gross assumed net for program, default is 1}

\item{NTG.eval}{is the percentage of gross evaluated net for program, default is 1}

\item{altparamnames}{is an optional vector of names for the parameters, default is the names in df[,1]; if supplied, ensure same order as df}

\item{output}{is an optional request for table, takes "none","gross","net","hybrid","all", default is "all". Note that "none" and "gross" are the same table for additive parameters.}
}
\value{
if output is "all", returns a list of 4 dataframes. Else, if output is "gross","net", or "hybrid", returns a dataframe with columns variable (character), given, total, calc, decrease, increase, and base (numeric). If output is "none" returns the same, less "calc".
}
\description{
addwaterfallPrep() : a function that takes the data on the reported and evaluated savings values and returns the average permuted values of the additive factors and prepare a data frame that is ready for the waterfall plot in R or Excel
}
\examples{
 addrawparamdf <- data.frame( # lighting example
                         params = c("A","B","C"),
                         value = c(10, -20, 25),
                         stringsAsFactors = FALSE
                        )
 #assume gross.report is 200, NTG.report is 0.6 & NTG.eval is 0.8
 addwaterfallPrep(rawparamdf, 200, 0.6, 0.8)

 #To just get the net permute table and rename the variables, example
 addwaterfallPrep(rawparamdf, 200, 0.6, 0.8,
 altparamnames=c("Prog A", "Prog B", "Prog C"),
 output = "net")

}

