% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/waterfallPlot.R
\name{waterfallPlot}
\alias{waterfallPlot}
\title{Makes a waterfall plot}
\usage{
waterfallPlot(df, palette = c("#d7191c", "#fdae61", "#abdda4", "#2b83ba"),
  xlab = "", ylab = "", offset = 0.3)
}
\arguments{
\item{df}{a dataframe with columns 'variable' (character), 'total' (numeric),  'base' (numeric), 'increase' (numeric), and 'decrease' (numeric). can have additional columns, ignored. Assumed to have starting value as first row and ending value as last row.}

\item{palette,}{default is c("#d7191c","#fdae61","#abdda4","#2b83ba")}

\item{xlab,}{default is ""}

\item{ylab,}{default is ""}

\item{offset,}{default is "0.3"}
}
\value{
a ggplot2 object
}
\description{
waterfallPlot() : a function that takes a dataframe and creates a waterfall plot as a ggplot2 object which can be further modified
}
\examples{
rrdf <- data.frame( # made up example
         variable = c("Start","Factor 1","Factor 2","Factor 3","End"),
         total = c(100, rep(NA, 3), 75),
         base = c(NA, 75, 50, 50,NA),
         increase = c(NA, 0, 0, 50, NA),
         decrease = c(NA, 25, 25, 0, NA))
 waterfallPlot(rrdf)
 # With another color palette. Note that totals stay grey.
 waterfallPlot(rrdf, palette=c("green","purple"))

# without intermediate tables, print the gross.permute plot
rawparamdf <- data.frame( # lighting example
                         params = c("ISR","deltaWatts","HOU","x"),
                         value = c(0.5, 0.7, 1.2, 1.5),
                         stringsAsFactors = FALSE
                        )
waterfallPlot(waterfallPrep(rawparamdf, 200, .6, .8, output="gross"))
}

