% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/waterfallPrep.R
\name{waterfallPrep}
\alias{waterfallPrep}
\title{Prepares data for a waterfall plot}
\usage{
waterfallPrep(df, gross.report, NTG.report = 1, NTG.eval = 1,
  altparamnames = NULL, output = "all")
}
\arguments{
\item{df}{a dataframe with columns 'params' (character) and 'value' (numeric); column names are immaterial. This is the order dependent value.}

\item{gross.report}{is the reported gross savings for the program}

\item{NTG.report}{is the percentage of gross assumed net for program, default is 1}

\item{NTG.eval}{is the percentage of gross evaluated net for program, default is 1}

\item{altparamnames}{is an optional vector of names for the parameters, default is the names in df[,1]; if supplied, ensure same order as df}

\item{output}{is an optional request for table, takes "none","gross","net","all", default is "all"}
}
\value{
a data frame object
}
\description{
waterfallPrep() : a function that takes the data on the reported and evaluated savings values and returns the average permuted values of the multiplicative factors and prepare a data frame that is ready for the waterfall plot in R or Excel
}
\examples{
rawparamdf <- data.frame( # lighting example
                         params = c("ISR","deltaWatts","HOU","x"),
                         value = c(0.5, 0.7, 1.2, 1.5),
                         stringsAsFactors = FALSE
                        )
 #assume gross.report is 200, NTG.report is 0.6 & NTG.eval is 0.8
 waterfallPrep(rawparamdf, 200, 0.6, 0.8)

 #To just get the net permute table and rename the variables, example
 waterfallPrep(rawparamdf, 200, 0.6, 0.8,
 altparamnames=c("ISR", "Dif Watts", "Daily Hours", "Control"),
 output = "net")
}

