% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/wParamPermute.R
\name{wParamPermute}
\alias{wParamPermute}
\title{The internal parameter permutation function}
\usage{
wParamPermute(param.names, values, n.cores = 2)
}
\arguments{
\item{param.names}{names of multiplicative parameters to permute (character vector)}

\item{values}{values of multiplicative parameters to permute (numeric vector)}

\item{n.cores}{number of cores for doParallel (numeric), default = 2}
}
\description{
This function is internal and used for creating order independent params
}
\examples{
wParamPermute(param.names = c("one", "hundred", "five"), values = c(1, 100, 5))
#Performance may be slow with large numbers of parameters. sytem.time elapsed is
#3 params ~ 0.011 | 5 params ~ 0.042 | 7 params ~ 1.741 | 8 params ~ 39.224
}
\keyword{internal}

